refactored to pandas
